Remove rocblas_wrapper.h
